+++ CHANGELOG.md
@@ -5,6 +5,13 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- 3 new programs
+	1. `work`
+	2. `short-break`
+	3. `long-break`
+### Changed
+- `PROGRAMS.md` updated
 ## [0.8] - 2023-08-07
 ### Added
 - Logo